Menu, articles (feed), photos
Added Main menu (Dan, please add .active to the ul that is selected in
the CMS). Added the template and styling for an article, as well as
photo with a good lightbox. Photos should be stored on our server in a
folder and randomly dumped in between the articles. Articles don’t have
random colours since there will be enough colours from the photos from
a visitor to perceive.